Make application a `template <class executor_t>` as Kevin suggested.
