+++ include/appbase/application.hpp
@@ -19,4 +19,8 @@
 
 #include <appbase/default_executor.hpp>
 
+namespace appbase {
+using application = application_t<default_executor>;
+}
+
 #include <appbase/application_instance.hpp>
